confirm-rx-usage: Verify that confirm-rx is properly implemented
When sending packets to a remote host the confirm-rx bit must be set on
some packets, to implement flow control and limit the resource usage on
the remote.

This commit implements a test that sends packets to a range of remote
ports and verifies that the maximum number of messages between two
messages with confirm-rx set is limited.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>