make: Introduce make ramdisk
Simplify generation of a ramdisk for qrtr-tests, by providing a make
target for installing and overlaying the tests onto an existing ramdisk.

This has been tested with initramfs-bootrr*.cpio.gz from
http://snapshots.linaro.org/96boards/dragonboard410c/linaro/openembedded/rocko/latest/rpb/

Download the file and run:
  make ramdisk RAMDISK_TEMPLATE=initramfs-bootrr*cpio.gz

Note that this will set CC to aarch64-linux-gnu-gcc, but won't rebuild
any up to date executables, so run make clean first.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>